@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 30
     defaultConfig {
         vectorDrawables.useSupportLibrary = true
         applicationId "com.curtesmalteser.jsoncurrencies"
-        minSdkVersion 16
-        targetSdkVersion 27
+        minSdkVersion 23
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -17,11 +17,12 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    dataBinding.enabled = true
-    buildToolsVersion '27.0.2'
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
+    }
+    buildFeatures {
+        dataBinding true
     }
 }
 
@@ -31,24 +32,24 @@ ext {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation "com.android.support:appcompat-v7:$supportLibraryVersion"
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation "com.android.support:support-annotations:$supportLibraryVersion"
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test:rules:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.annotation:annotation:1.2.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
+    androidTestImplementation 'androidx.test:rules:1.3.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
     // Add the RecyclerView dependency
-    implementation "com.android.support:recyclerview-v7:$supportLibraryVersion"
+    implementation 'androidx.recyclerview:recyclerview:1.2.0'
     // Add the Preferences dependency
-    implementation "com.android.support:preference-v7:$supportLibraryVersion"
+    implementation 'androidx.preference:preference:1.1.1'
     // Add the dependency LiveData and ViewModel
-    implementation 'android.arch.lifecycle:extensions:1.0.0'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
     // Add ROOM dependencies
-    implementation 'android.arch.persistence.room:runtime:1.0.0'
-    annotationProcessor "android.arch.persistence.room:compiler:1.0.0"
+    implementation 'androidx.room:room-runtime:2.3.0'
+    annotationProcessor 'androidx.room:room-compiler:2.3.0'
     // Lifecycles
-    annotationProcessor "android.arch.lifecycle:compiler:1.0.0"
+    annotationProcessor 'androidx.lifecycle:lifecycle-common-java8:2.3.1'
 
     // Firebase JobDispatcher
     implementation 'com.firebase:firebase-jobdispatcher:0.5.2'
