@@ -2,8 +2,8 @@ package com.curtesmalteser.jsoncurrencies.activity;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.curtesmalteser.jsoncurrencies.R;
 import com.curtesmalteser.jsoncurrencies.fragment.MasterListFragment;
