@@ -1,9 +1,9 @@
 package com.curtesmalteser.jsoncurrencies.db;
 
 import android.app.Application;
-import android.arch.lifecycle.AndroidViewModel;
-import android.arch.lifecycle.LiveData;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.annotation.NonNull;
 
 import com.curtesmalteser.jsoncurrencies.model.CurrenciesModel;
 
