@@ -3,10 +3,9 @@ package com.curtesmalteser.jsoncurrencies.widget;
 import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
-import android.os.AsyncTask;
 import android.os.Binder;
-import android.support.annotation.Nullable;
-import android.support.v4.content.AsyncTaskLoader;
+import androidx.annotation.Nullable;
+import androidx.loader.content.AsyncTaskLoader;
 import android.util.Log;
 import android.widget.RemoteViews;
 import android.widget.RemoteViewsService;
