@@ -1,11 +1,11 @@
 package com.curtesmalteser.jsoncurrencies.db;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Delete;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
-import android.arch.persistence.room.Update;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Delete;
+import androidx.room.Insert;
+import androidx.room.Query;
+import androidx.room.Update;
 import android.database.Cursor;
 
 import com.curtesmalteser.jsoncurrencies.model.CurrenciesModel;
@@ -13,7 +13,7 @@ import com.curtesmalteser.jsoncurrencies.model.CurrenciesModel;
 import java.util.ArrayList;
 import java.util.List;
 
-import static android.arch.persistence.room.OnConflictStrategy.REPLACE;
+import static androidx.room.OnConflictStrategy.REPLACE;
 
 /**
  * Created by António "Curtes Malteser" Bastião on 14/12/2017.
